Added Fixes
