+++ handraiser-app/src/index.js
@@ -1,32 +1,10 @@
-import React, { useState } from "react";
+import React from "react";
 import ReactDOM from "react-dom";
-import GoogleLogin from "react-google-login";
-import Login from "./components/Login/Login";
-function App() {
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [url, setUrl] = useState("");
-  const responseGoogle = response => {
-    console.log(response);
-    setName(response.profileObj.name);
-    setEmail(response.profileObj.email);
-    setUrl(response.profileObj.imageUrl);
-  };
-  return (
-    <div className="App">
-      <h1>Login with Google</h1>
-      <h2>Welcome:{name}</h2>
-      <h2>Email:{email}</h2>
-      <img src={url} alt={name} />
-      <GoogleLogin
-        clientId="239954847882-ilomcrsuv3b0oke6tsbl7ofajjb11nkl.apps.googleusercontent.com"
-        buttonText="Login"
-        onSuccess={responseGoogle}
-        onFailure={responseGoogle}
-        cookiePolicy={"single_host_origin"}
-      />
-    </div>
-  );
-}
-const rootElement = document.getElementById("root");
-ReactDOM.render(<Login />, rootElement);
+import "./index.css";
+import App from "./App";
+import * as serviceWorker from "./serviceWorker";
+ReactDOM.render(<App />, document.getElementById("root"));
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();
